+++ CHANGELOG.md
@@ -51,6 +51,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Config generation: rename `Pleroma.Upload.Filter.ExifTool` to `Pleroma.Upload.Filter.Exiftool`.
 - S3 Uploads with Elixir 1.11.
+- Fixed Emoji Reaction activity filtering from blocked and muted accounts
 - Mix task pleroma.user delete_activities for source installations.
 
 ## [2.2.0] - 2020-11-12